made DB URL in config and share it across added helm confgimap entries and values.yaml entries added more logs and await added transaction boundary for delete across tables.
