+++ backend/src/v1/prisma/prisma-client.ts
@@ -1,17 +1,8 @@
 import { logger } from '../../logger';
 import { Prisma, PrismaClient } from '@prisma/client';
+import { config } from '../../config';
 
-const DB_HOST = process.env.POSTGRESQL_HOST || 'localhost';
-const DB_USER = process.env.POSTGRESQL_USER || 'postgres';
-const DB_PWD = encodeURIComponent(
-  process.env.POSTGRESQL_PASSWORD || 'postgres',
-);
-const DB_PORT = process.env.POSTGRESQL_PORT || 5432;
-const DB_NAME = process.env.POSTGRESQL_DATABASE || 'postgres';
-const DB_SCHEMA = process.env.DB_SCHEMA || 'pay_transparency';
 
-const datasourceUrl = `postgresql://${DB_USER}:${DB_PWD}@${DB_HOST}:${DB_PORT}/${DB_NAME}?schema=${DB_SCHEMA}&connection_limit=5`;
-logger.silly(`Connecting to ${datasourceUrl}`);
 const prisma: PrismaClient<
   Prisma.PrismaClientOptions,
   'query' | 'info' | 'warn' | 'error'
@@ -23,7 +14,7 @@ const prisma: PrismaClient<
     { emit: 'stdout', level: 'error' },
   ],
   errorFormat: 'pretty',
-  datasourceUrl: datasourceUrl,
+  datasourceUrl: config.get('server:databaseUrl'),
 });
 prisma.$on('query', (e) => {
   logger.debug(
