+++ charts/fin-pay-transparency/templates/configmap.yaml
@@ -8,4 +8,5 @@ data:
   UPLOAD_FILE_MAX_SIZE: {{ .Values.global.config.upload_file_max_size | quote }}
   TEMPLATE_PATH: {{ .Values.global.config.template_path | quote }}
   REPORT_EDIT_DURATION_IN_DAYS: {{ .Values.global.config.report_edit_duration_in_days | quote }}
-
+  DELETE_DRAFT_REPORT_CRON_CRONTIME: {{ .Values.global.config.delete_draft_report_cron_crontime | quote }}
+  DELETE_DRAFT_REPORT_CRON_TIMEZONE: {{ .Values.global.config.delete_draft_report_cron_timezone | quote }}
