+++ charts/fin-pay-transparency/values.yaml
@@ -28,6 +28,8 @@ global:
     upload_file_max_size: '8000000'
     report_edit_duration_in_days: '30'
     template_path: './dist/templates'
+    delete_draft_report_cron_crontime: '0 0 0 * * *' # 12:00 AM PST/PDT
+    delete_draft_report_cron_timezone: 'America/Vancouver'
   crunchyEnabled: true
 backend:
   enabled: true
